services/speech_to_text/requirements.txt
Flask
google-cloud-speech
google-cloud-storage